Changes to onSelect for ColorPicker,  SmallBoxList Pass id in onSelect rather than full object
